% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_binomial.R
\name{ci_binomial}
\alias{ci_binomial}
\title{Estimate confidence intervals for binomial distributions}
\usage{
ci_binomial(
  data,
  conf = 0.95,
  summarized = FALSE,
  phat = NULL,
  fpc = FALSE,
  n = NULL,
  N = NULL,
  method = "agresti.coull"
)
}
\arguments{
\item{data}{A vector of binary data.  Required if \code{summarized = FALSE}.}

\item{conf}{Level of confidence \verb{1 - P(type I error)}.}

\item{summarized}{Logical; indicate whether raw data or summary stats are to be used.}

\item{phat}{Estimate of π. Required if \code{summarized = TRUE}.}

\item{fpc}{Logical. Indicates whether finite population corrections should be
used. If \code{fpc = TRUE} then \code{N} must be specified. Finite population corrections
are not possible for \code{method = "exact"} or \code{method = "score"}.}

\item{n}{Sample size. Required if \code{summarized = TRUE}.}

\item{N}{Population size. Required if \code{fpc = TRUE}.}

\item{method}{Type of method to be used in confidence interval calculations,
\code{method ="agresti.coull"} is the default. Other procedures include
\code{method = "asymptotic"} which provides the conventional normal (Wald)
approximation, \code{method = "score"}, \code{method = "LR"}, and \code{method = "exact"} (see
Details below). Partial names can be used. The \code{"exact"} method cannot be
implemented if \code{summarized=TRUE}.}
}
\value{
Returns a list of class = "ci".
}
\description{
This function is an almost exact copy of the \code{asbio::ci.p()} function. The
only differences are that (1) the \code{plot} argument and associated functionality
have been removed, and (2) the output \code{head} has been slighlty altered.
}
\details{
See \code{?asbio::ci.p} for details.
}
\author{
See \code{?asbio::ci.p}
}
\keyword{internal}
