% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_binomial.R
\name{col_binomial}
\alias{col_binomial}
\alias{is_col_binomial}
\title{Binomial Column}
\usage{
col_binomial(
  successes = integer(),
  sample = integer(),
  ci_error = 0.05,
  population = Inf
)

is_col_binomial(x)
}
\arguments{
\item{successes}{A numeric vector, the number of successes}

\item{sample}{A numeric vector, the number of Bernoulli trials}

\item{ci_error}{A numeric vector, the error to be used for calculating confidence intervals}

\item{population}{A numeric vector, the number of individuals in the population to be used for calculating confidence intervals}

\item{x}{An object to test}
}
\value{
An S3 vector of class \code{projectable_col_binomial}
}
\description{
Create, or test for, objects of type \code{projectable_col_binomial}.
}
\details{
Confidence intervals are calculated using the Agresti-Coull method. If the
population size is not \code{Inf} a finite population correction will be applied.
}
\examples{

b_trials <- stats::rbinom(1000, 1, 0.5)
col_binomial(sum(b_trials), length(b_trials))

}
