% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_binomial.R
\name{col_binomial_vec}
\alias{col_binomial_vec}
\title{Calculate binomial proportions from a vector}
\usage{
col_binomial_vec(
  x,
  success_value = 1,
  range = c(0, 1),
  ci_error = 0.05,
  population = Inf,
  method = "agresti.coull"
)
}
\arguments{
\item{x}{A numeric or integer vector containing the data. Can include NA values.}

\item{success_value}{The value(s) that count as "success". Can be a single value
or a vector of values. Default is 1.}

\item{range}{The range of values to include in the calculation. Values outside
this range are excluded (treated as NA). Default is c(0, 1). Set to NULL to
include all non-NA values.}

\item{ci_error}{A numeric vector, the error to be used for calculating
confidence intervals. Default is 0.05.}

\item{population}{A numeric vector, the number of individuals in the
population to be used for calculating confidence intervals. Default is Inf.}

\item{method}{The name of a method to be passed through to \code{asbio::ci.p()} for
parameter estimation. The default is "agresti.coull", but other options
include "asymptotic", "score", "LR" and "exact". See \code{asbio::ci.p()} for details.}
}
\value{
An S3 vector of class \code{projectable_col_binomial}
}
\description{
A wrapper function for \code{col_binomial()} that takes a single vector and
automatically calculates the number of successes and trials. This is
particularly useful when working with data that contains missing values.
}
\examples{
# Calculate proportion of 1s in a 0-1 vector with missing data
x <- c(0, 1, 1, 0, NA, 1, 0, 1)
col_binomial_vec(x)

# Calculate proportion of 2s in a 1-2 vector
y <- c(1, 2, 2, 1, NA, 2, 1, 2)
col_binomial_vec(y, success_value = 2, range = c(1, 2))

# Calculate proportion with multiple success values
z <- c(1, 2, 3, 4, 5, NA, 2, 3)
col_binomial_vec(z, success_value = c(2, 3), range = c(1, 5))

}
