% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_row.R
\name{col_row}
\alias{col_row}
\alias{is_col_row}
\title{Create and test for row-label columns}
\usage{
col_row(x = character())

is_col_row(x)
}
\arguments{
\item{x}{An atomic vector to test or to turn into a \code{col_row}}
}
\value{
An S3 vector of class \code{projectable_col_row}
}
\description{
A dataframe representing a table will have one or more columns which contain
the labels of the rows. By designating them as such using the \code{col_row()}
function, we can make sure they are treated differently from the non-row
columns. In particular, any \code{col_row} columns will be retained through
projections.
}
\examples{


library(dplyr)
mtcars \%>\%
  # Group by a `col_row`
  group_by(vs = col_row(vs)) \%>\%
  summarise(
    Manual = col_freq(am \%in\% 1, am \%in\% 0:1),
    Auto = col_freq(am \%in\% 0, am \%in\% 0:1)
  ) \%>\%
  # We don't have to worry about applying a special shadow to the `col_row`
  prj_shadow_all("{signif(p, 2)} ({n})") \%>\%
  prj_gt(rowgroup_col = NULL, rowname_col = NULL)

}
