% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encol.R
\name{encol_binomial}
\alias{encol_binomial}
\title{encol_binomial}
\usage{
encol_binomial(success, sample, ci_error = 0.05, population = Inf)
}
\arguments{
\item{success}{An unquoted expression that evaluates to a logical, defining successes; alternatively an integer representing the number of successes}

\item{sample}{An unquoted expression that evaluates to a logical, defining the sample size; alternatively an integer representing the sample size}

\item{ci_error}{A scalar number representing the error, which is used to calculate confidence intervals.}

\item{population}{An unquoted expression that evaluates to a logical, defining the population size; alternatively an number representing the population size}
}
\value{
A quoted call to \code{col_binomial()}
}
\description{
Generate a quoted call to \code{col_binomial()}, typically to be evaluated within the context of a call to \code{set_table()}.
}
\examples{
# On its own
encol_binomial(vs \%in\% 1, vs \%in\% 0:1)

# Within the context of `set_table()`
set_table(
  .data = mtcars,
  .rows = cyl,
  .cols = list(
    vshaped = encol_binomial(vs \%in\% 1, vs \%in\% 0:1),
    non_vshaped = encol_binomial(vs \%in\% 0, vs \%in\% 0:1)
  )
)
}
