% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_table.R
\name{encol_freq}
\alias{encol_freq}
\title{encol_freq}
\usage{
encol_freq(little_n, big_n)
}
\arguments{
\item{little_n}{an unquoted expression which evaluates to a logical, defining the count}

\item{big_n}{an unquoted expression which evaluates to a logical, defining the base count}
}
\value{

}
\description{
encol_freq
}
\examples{



}
