% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face_value.R
\name{face_value}
\alias{face_value}
\title{Get face values}
\usage{
face_value(x)
}
\arguments{
\item{x}{A \code{projectable_col}}
}
\value{
An atomic vector representing the face value of \code{x}
}
\description{
Each \code{projectable_col} stores a face value alongside a range of metadata.
This function extracts that value.
}
\details{
If \code{x} is not a \code{projectable_col}, \code{face_value()} simply returns \code{x} unaltered.
}
\examples{

face_value(col_freq(1:5, 6:10, summarised = TRUE))

face_value(1:5)
}
