% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{is_col}
\alias{is_col}
\alias{col_shadow}
\alias{col_shadow<-}
\title{Test, get and set cols and their attributes}
\usage{
is_col(x)

col_shadow(x)

col_shadow(x) <- value
}
\arguments{
\item{x}{An object to test or to check the attributes of}

\item{value}{A value to set the attribute to}
}
\description{
There are various subclasses of \code{projectable_col} including the \code{col_freq} and
\code{col_binomial}. These functions provide a high-level interface to those subclasses.
}
