% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_tbl.R
\name{list_to_dataframe}
\alias{list_to_dataframe}
\title{Convert a nested list to a tibble}
\usage{
list_to_dataframe(x, i = 0)
}
\arguments{
\item{x}{A nested list (i.e. a list of lists of lists of...)}

\item{i}{A starting index to suffix \code{row_} columns with}
}
\value{
A tibble
}
\description{
The innermost list is bound together column-wise; the other lists are bound
together row-wise. The names of the innermost list become the column names,
and proceeding outwards the names of successive lists are stored within
columns of the output tibble starting as \code{row_i}, then \code{row_i+1}, etc.
}
