% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{new_col}
\alias{new_col}
\title{Create new \code{projectable_col}vectors}
\usage{
new_col(..., shadow = character(), class = character())
}
\arguments{
\item{...}{Fields to store in the \code{projectable_col}}

\item{shadow}{A shadow to initialise the \code{projectable_col} with}

\item{class}{A character vector giving the \code{projectable_col} subclass.}
}
\value{
A \code{projectable_col} vector with subclass \verb{projectable_col_*} where
\code{*} is given by whatever was passed in as the \code{class} argument.
}
\description{
To create a new class of \code{projectable_col} vectors, one simply needs to
define a \verb{col_*()} helper function. This can be done with the \code{new_col()}
constructor function. The first field passed into the \code{new_col()} function
via the \code{...} argument will be taken as the face value. This can be overriden
by explicitly defining a \code{face_value()} method. The subclass of the
\code{projectable_col} created in this way will be \verb{projectable_col_*} where \code{*}
is replaced by whatever is passed in as the \code{class} argument.
}
\examples{

# Define a `col_student` helper function:
col_student <- function(name = character(), age = integer(), grade = double()) {
  new_col(name = name, age = age, grade = grade, class = "student")
}
col_student(c("Kinto", "Pinto"), c(25, 100), c(99, 100))

}
