% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_project.R
\name{prj_project}
\alias{prj_project}
\alias{prj_gt}
\alias{prj_flex}
\title{Project metadata-rich tables}
\usage{
prj_project(.data, .cols = list(), .digits = getOption("prj_digits"))

prj_gt(
  .data,
  .cols = list(),
  .digits = getOption("prj_digits"),
  rowgroup_col = "row_spanner",
  rowname_col = "rows",
  ...
)

prj_flex(.data, .cols = list(), .digits = getOption("prj_digits"))
}
\arguments{
\item{.data}{A dataframe, ideally one containing \code{projectable_col}s}

\item{.cols}{A named list. Each name should  be the name of a column in
\code{.data}; each value should be a named character vector containing glue-like
specifications for the output columns.}

\item{.digits}{A number representing the number of digits to round each
numeric value to. If \code{NULL} no rounding will be performed}

\item{rowgroup_col}{The name of a column in \code{.data} to group rows by; if
\code{NULL} no grouping will be used.}

\item{rowname_col}{The name of a column in \code{.data} to take as the row labels;
if \code{NULL} no row labels will be applied}

\item{...}{Additional arguments to pass on to \code{gt::gt()}}
}
\value{
a \code{projection} object
}
\description{
The \code{prj_project()} function is designed to take in a dataframe made up of
\code{projectable_col}s, and to 'project' it into an ordinary dataframe as per the
instructions provided in the \code{shadow} attribute of each column. The
\code{prj_gt()} function does the same thing, but initialises a \code{gt} object for
display.
}
\details{
The \code{shadow} attribute of each column can be set via the \code{.cols} argument of
\code{prj_project()} or by using the \code{prj_shadow()} helper function.

The \code{projection} output will also come attached with metadata which keeps
track of which columns in the output belong to which columns in the input.
}
\examples{
# Create a table made up of `projectable_col`s
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = list(Automatic = am \%in\% 0, Manual = am \%in\% 1),
)

my_tbl <- prj_tbl_cols(
  .data = my_tbl,
  `V-Shaped` = col_freq(n = vs \%in\% 1, N = vs \%in\% 0:1),
  `Not V-shaped` = col_freq(n = vs \%in\% 0, N = vs \%in\% 0:1)
)

my_tbl <- prj_tbl_summarise(.data = my_tbl)

# Project it back into an ordinary dataframe
prj_project(my_tbl, list(
  `V-Shaped` = "{signif(p, 2)} ({n})",
  `Not V-shaped` = "{signif(p, 2)} ({n})"
))

# Produce a `gt` display object
out <- prj_gt(my_tbl, list(
  `V-Shaped` = "{signif(p, 2)} ({n})",
  `Not V-shaped` = "{signif(p, 2)} ({n})"
))

#' # Produce a `flextable` display object
out <- prj_flex(my_tbl, list(
  `V-Shaped` = "{signif(p, 2)} ({n})",
  `Not V-shaped` = "{signif(p, 2)} ({n})"
))

}
