% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_shadow.R
\name{prj_shadow}
\alias{prj_shadow}
\title{Specify projections for multiple columns}
\usage{
prj_shadow(.data, ..., .shadow)
}
\arguments{
\item{.data}{A dataframe, ideally one containing \code{projectabel_col}s}

\item{...}{\href{https://tidyselect.r-lib.org/articles/syntax.html}{\code{tidy-select}}
One or more unquoted expressions separated by commas. Variable names can be
used as if they were positions in the data frame, so expressions like \code{x:y}
can be used to select a range of variables.}

\item{.shadow}{A character vector containing glue-like specifications to
which to set the \code{shadow} attribute of the relevant columns.}
}
\value{
A dataframe, the columns of which have updated \code{shadow} attributes.
}
\description{
The transition from a metadata-rich dataframe consisting of
\code{projectable_col}s to an ordinary dataframe that is fit for presentation is
guided by the \code{shadow} attributes attached to each column of the initial
dataframe. The \code{prj_shadow()} function is designed to make it easy to
specify those \code{shadow} attributes for multiple columns at once.
}
\examples{
# Create a table made up of `projectable_col`s
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = list(Automatic = am \%in\% 0, Manual = am \%in\% 1),
)

my_tbl <- prj_tbl_cols(
  .data = my_tbl,
  `V-Shaped` = col_freq(n = vs \%in\% 1, N = vs \%in\% 0:1),
  `Not V-shaped` = col_freq(n = vs \%in\% 0, N = vs \%in\% 0:1)
)

my_tbl <- prj_tbl_summarise(.data = my_tbl)

# Update the `shadow` attributes
shadow <- c(Count = "{n}", Proportion = "{signif(p, 2)}")
my_tbl <- prj_shadow(my_tbl, where(is_col_freq), .shadow = shadow)

# Project it back into an ordinary dataframe
prj_project(my_tbl)

}
