% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_shadow.R
\name{prj_shadow_all}
\alias{prj_shadow_all}
\alias{prj_shadow_if}
\alias{prj_shadow_at}
\title{Specify projections for multiple columns}
\usage{
prj_shadow_all(.data, .shadow)

prj_shadow_if(.data, .predicate, .shadow)

prj_shadow_at(.data, .vars, .shadow)
}
\arguments{
\item{.data}{A dataframe, ideally one containing \code{projectabel_col}s}

\item{.shadow}{A character vector containing glue-like specifications to
which to set the \code{shadow} attribute of the relevant columns.}

\item{.predicate}{A predicate function to be applied to the columns or a
logical vector. The variables for which \code{.predicate} is or returns TRUE
have their \code{shadow} updated}

\item{.vars}{A character vector containing the names of columns in \code{.data} of
which to update the \code{shadow} attribute}
}
\value{
A dataframe, the columns of which have updated \code{shadow} attributes.
}
\description{
The transition from a metadata-rich dataframe consisting of
\code{projectable_col}s to an ordinary dataframe that is fit for presentation is
guided by the \code{shadow} attributes attached to each column of the initial
dataframe. The \verb{prj_shadow_*()} functions are designed to make it easy to
specify those \code{shadow} attributes for multiple columns at once.
}
\examples{
# Create a table made up of `projectable_col`s
my_tbl <- set_table(
  .data = mtcars,
  .rows = list(
    Cylinders = cyl
  ),
  .cols = list(
    vshaped = col_freq(n = vs \%in\% 1, N = vs \%in\% 0:1),
    not_vshaped = col_freq(n = vs \%in\% 0, N = vs \%in\% 0:1)
  )
)

# Update the `shadow` attributes
my_tbl <- prj_shadow_if(my_tbl, is_col_freq(.), c(Count = "{n}", Proportion = "{signif(p, 2)}"))

# Project it back into an ordinary dataframe
prj_table(my_tbl)

}
