% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prj_tbl.R
\name{prj_tbl_summarise}
\alias{prj_tbl_summarise}
\alias{prj_tbl_rows}
\alias{prj_tbl_cols}
\title{prj_tbl}
\usage{
prj_tbl_summarise(.data)

prj_tbl_rows(.data, ...)

prj_tbl_cols(.data, ...)
}
\arguments{
\item{.data}{A dataframe.}

\item{...}{A column or row specification. Row specifications may be supplied
as symbols/expressions or lists of symbols/expressions. Column specifications
may be supplied as expressions only.}
}
\value{
A tibble
}
\description{
Build a summary table with \code{prj_tbl_summarise()} as per column and row
specifications provided via \code{prj_tbl_cols()} and \code{prj_table_rows()}
respectively.
}
\examples{
# Create a table made up of `projectable_col`s
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = list(Automatic = am \%in\% 0, Manual = am \%in\% 1),
)

my_tbl <- prj_tbl_cols(
  .data = my_tbl,
  `V-Shaped` = col_freq(n = vs \%in\% 1, N = vs \%in\% 0:1),
  `Not V-shaped` = col_freq(n = vs \%in\% 0, N = vs \%in\% 0:1)
)

my_tbl <- prj_tbl_summarise(.data = my_tbl)

}
