% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectable-package.R
\docType{package}
\name{projectable-package}
\alias{projectable}
\alias{projectable-package}
\title{projectable: Metadata-Rich Tables}
\description{
Produce table-like objects made up of special metadata-rich
    column vectors, and project them into two dimensions.
}
\author{
\strong{Maintainer}: Kinto Behr \email{kinto.behr@srcentre.com.au}

Other contributors:
\itemize{
  \item The Social Research Centre [copyright holder]
}

}
\keyword{internal}
