% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_freq.R
\name{projectable-vctrs}
\alias{projectable-vctrs}
\alias{new_col_freq}
\title{Internal vctrs methods}
\usage{
new_col_freq(little_n = double(), big_n = double(), proportion = double())
}
\description{
Internal vctrs methods
}
\keyword{internal}
