% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{public_formats}
\alias{public_formats}
\alias{public_num}
\alias{public_ci}
\title{Format numbers for public consumption}
\usage{
public_num(x, N, digits = 1, min_N = 25)

public_ci(x, N, ci_lower = NA, ci_upper = NA, digits = 1, min_N = 25)
}
\arguments{
\item{x}{A numeric vector, typically the 'face value' of a projectable \code{col}}

\item{N}{A numeric vector, the base count}

\item{digits}{A number, the number of digits to values to}

\item{min_N}{A number, the minimum N required for values to be reportable}

\item{ci_lower}{A numeric vector, the lower bound of the confidence interval}

\item{ci_upper}{A numeric vector, the upper bound of the confidence interval}
}
\value{
A character vector
}
\description{
In public reports, cells are only reported if the N count is greater than
some minimum number (usually 25). In addition, we:
\itemize{
\item Round figures to some number of digits, retaining trailing 0s,
e.g. 25 -> 25.0
\item Display confidence intervals in brackets beside estimates,
e.g. '25.0 (24.5, 25.5)'
}
}
\examples{

# Present a formatted number with suppressions applied:
x <- projectable::col_freq(1:5, 1:5*10, summarised = TRUE)
x <- projectable::prj_project_col(x)
public_num(x$p, x$N, digits = 2)

# Present a formatted number and its confidence interval with suppressions applied
x <- projectable::col_binomial(1:5, 1:5*10, summarised = TRUE)
x <- projectable::prj_project_col(x)
public_ci(x$p, x$N, x$ci_lower, x$ci_upper)

}
