% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public.R
\name{rounding}
\alias{rounding}
\alias{dec_dig3}
\title{Rounding}
\usage{
dec_dig3(x, n, na_replacement = "")
}
\arguments{
\item{x}{A numeric vector}

\item{n}{An integer, how many digits to round to}

\item{na_replacement}{A string, what to replace NAs with}
}
\value{
A character vector
}
\description{
\code{dec_dig3()} is essentially a wrapper for \code{srcutils::round2()}. It formats
numbers by:
\itemize{
\item Rounding them using \code{srcutils::round2()}
\item Inserting commas as thousands separators (e.g. as in "10,000")
\item Retaining trailing 0s (e.g. as in "10.0")
}
}
