% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_col.R
\name{spec_col_freq}
\alias{spec_col_freq}
\title{Frequency cross-tabulations}
\usage{
spec_col_freq(.var, .vals, .type = "row")
}
\arguments{
\item{.var}{An unquoted variable name}

\item{.vals}{A vector of values \code{.var} might take}

\item{.type}{A string designating the type of frequency to calculate: "row", "col" or "cell"}
}
\value{
A list of expressions to be passed into \code{prj_tbl_cols()}
}
\description{
Conveniently generate frequency column specifications by passing through an
unquoted variable name and a set of valid values. This specification can then
be spliced into \code{prj_tbl_cols()}.
}
\examples{
# Specify the rows
my_tbl <- prj_tbl_rows(
  .data = mtcars,
  Cylinders = cyl,
  Transmission = am,
)

# Specify the columns
my_tbl1 <- prj_tbl_cols(
  .data = my_tbl,
  !!!spec_col_freq(vs, 0:1)
)

# Summarise
prj_tbl_summarise(my_tbl1)

# Specify the columns and provide custom names
col_spec <- `names<-`(spec_col_freq(vs, 0:1), c("Not V-Shaped", "V-Shaped"))
my_tbl2 <- prj_tbl_cols(
  .data = my_tbl,
  !!!col_spec
)

# Summarise
prj_tbl_summarise(my_tbl2)
}
