% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_freq.R
\name{col_freq}
\alias{col_freq}
\alias{is_col_freq}
\title{Frequency Columns}
\usage{
col_freq(n = double(), N = double(), p = NULL)

is_col_freq(x)
}
\arguments{
\item{n}{A numeric vector of frequencies}

\item{N}{A numeric vector of base frequencies}

\item{p}{A numeric vector of proportions, defaults to the quotient of \code{n} and \code{N}}

\item{x}{An object to test}
}
\value{
An S3 vector of class \code{projectable_col_freq}
}
\description{
Create, or test for, objects of type \code{projectable_col_freq}.
}
\examples{
col_freq(1, 2, 1/2)

}
