% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_freq.R
\name{col_freq}
\alias{col_freq}
\title{Frequency Columns}
\usage{
col_freq(little_n = double(), big_n = double(), proportion = NULL)
}
\arguments{
\item{little_n}{A numeric vector of frequencies}

\item{big_n}{A numeric vector of base frequencies}

\item{proportion}{A numeric vector of proportions, defaults to the quotient of \code{little_n} and \code{big_n}}
}
\value{
An S3 vector of class \code{projectable_col_freq}
}
\description{
Frequency Columns
}
\examples{
col_freq(1, 2, 1/2)

}
