% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encol.R
\name{encol_freq}
\alias{encol_freq}
\title{encol_freq}
\usage{
encol_freq(little_n, big_n)
}
\arguments{
\item{little_n}{An unquoted expression which evaluates to a logical, defining the count; alternatively an integer representing the count}

\item{big_n}{An unquoted expression which evaluates to a logical, defining the base count; alternatively an integer representing the base count}
}
\value{
A quoted call to \code{col_freq()}
}
\description{
Generate a quoted call to \code{col_freq()}, typically to be evaluated within the context of a call to \code{set_table()}.
}
\examples{

# On its own
encol_freq(vs \%in\% 1, vs \%in\% 0:1)

# Within the context of `set_table()`
set_table(
  .data = mtcars,
  .rows = cyl,
  .cols = list(
    vshaped = encol_freq(vs \%in\% 1, vs \%in\% 0:1),
    non_vshaped = encol_freq(vs \%in\% 0, vs \%in\% 0:1)
  )
)

}
